fix(relay): force dynamic rendering on setup-gated routes
Cherry-pick of the squashed PR #68 (https://github.com/hanbin2007/AIChat/pull/68)
from v3 onto main. Verified deployed and working on ai.origenclub.cn — `/`,
`/dashboard`, and `/setup` now redirect setup-complete users to `/login`
instead of bouncing them back to the setup wizard.

## What's in this commit

Production fix:
- Mark `/` and `(admin)/layout` as `export const dynamic = "force-dynamic"`
  so the setup gate is re-evaluated per request.
- Split `/setup` into a server guard + client form (page.tsx + setup-form.tsx)
  and mark it dynamic too. The guard redirects to `/login` when setup is
  already complete, so the page can never trap a fully-initialized relay.

Sandbox tooling (also part of #68 — relay box SSH-over-WSS bring-up):
- scripts/server-setup-chisel.sh — installs chisel as a systemd unit on the
  relay box and patches /etc/caddy/Caddyfile to add a /_chisel/ route.
- scripts/sandbox-tunnel.sh — sandbox-side bring-up; one command after env
  secrets are set, gives `ssh rt` to the relay box from inside the sandbox.
- scripts/server-recon.sh — read-only inventory script for the relay host.
- docs/relay-server-setup.md — full topology, verify, rotate, revoke.
- CLAUDE.md — appends a "Relay server access" runbook.

CLAUDE.md auto-merged cleanly (main's architecture section is older than v3's
and stays untouched; the new section is appended at the bottom).
