+++ relay/src/app/(admin)/layout.tsx
@@ -2,6 +2,8 @@ import { redirect } from "next/navigation";
 import { readSession } from "@/lib/auth/session";
 import { settingsStore } from "@/lib/store/settings-store";
 
+export const dynamic = "force-dynamic";
+
 export default async function AdminLayout({ children }: { children: React.ReactNode }) {
   const setup = await settingsStore().isSetupComplete();
   if (!setup) redirect("/setup");
